+++ database/migrations/2025_08_22_223958_alter_exams_subject_id_nullable.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::table('exams', function (Blueprint $table) {
             // si la FK existe déjà et est NOT NULL, il faut souvent la drop/recreate
             try { $table->dropForeign(['subject_id']); } catch (\Throwable $e) {}
-            $table->unsignedBigInteger('subject_id')->nullable()->change();
+            $table->integer('subject_id')->nullable()->change();
             $table->foreign('subject_id')->references('id')->on('subjects')->nullOnDelete();
         });
     }
